+++ apps/examples/wc-next10/ButtonSizes.js
@@ -4,22 +4,18 @@ import { PieButton } from '@justeattakeaway/pie-button/dist/react';
 
 export default function ButtonSizes () {
     return (
-        <>
+        <div>
+            <h3>Button sizes</h3>
             <div>
-                <h3>Button sizes</h3>
-                <div>
-                    {
-                        buttonSizes
-                            .map((key, index) => {
-                                const size = buttonSizes[index];
-
-                                return (
-                                    <PieButton size={size} key={key}>{key.toLowerCase()}</PieButton>
-                                );
-                            })
-                    }
-                </div>
+                {buttonSizes.map((key, index) => {
+                    const size = buttonSizes[index];
+                    return (
+                        <PieButton size={size} key={key}>
+                            {key.toLowerCase()}
+                        </PieButton>
+                    );
+                })}
             </div>
-        </>
+        </div>
     );
 }