+++ apps/examples/wc-next10/Counter.js
@@ -8,17 +8,13 @@ export default function Counter () {
     const onDecrement = () => setCount(count - 1);
 
     return (
-        <>
-            <div>
-                <h3>Counter</h3>
-                <div className="flex-wrapper">
-                    <PieButton onClick={onDecrement}>decrement</PieButton>
-                    <div className="padding">
-                        Counter: { count }
-                    </div>
-                    <PieButton onClick={onIncrement}>increment</PieButton>
-                </div>
+        <div>
+            <h3>Counter</h3>
+            <div className="flex-wrapper">
+                <PieButton onClick={onDecrement}>decrement</PieButton>
+                <div className="padding">Counter: {count}</div>
+                <PieButton onClick={onIncrement}>increment</PieButton>
             </div>
-        </>
+        </div>
     );
 }