+++ apps/examples/wc-next13/.eslintrc.js
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
     extends: [
         require.resolve('@justeattakeaway/eslint-config-pie/strict'),
         'eslint:recommended',