+++ configs/pie-components-config/index.js
@@ -1,4 +1,2 @@
-/* eslint-disable import/extensions */
 export * from './playwright-lit-config.js';
 export * from './playwright-lit-visual-config.js';
-