+++ packages/components/pie-button/src/defs.ts
@@ -1,13 +1,35 @@
 export interface ButtonProps {
+    /**
+     * the size of the button.
+     * @default medium
+     */
     size: 'xsmall' | 'small-expressive' | 'small-productive' | 'medium' | 'large';
+    /**
+     * The html button type to use.
+     * @default submit
+     */
     type: 'submit' | 'button' | 'reset' | 'menu';
+    /**
+     * the variant of the button.
+     * @default primary
+     */
     variant: 'primary' | 'secondary' | 'outline' | 'ghost';
+    /**
+     * If `true`, the button will be disabled.
+     * @default false
+     */
+    disabled: boolean;
+    /**
+     * If `true`, the button will span the full width.
+     * @default false
+     */
+    isFullWidth: boolean;
 }
 
 /**
  * Button size variants
  */
-export const buttonSizes: ButtonProps['size'][] = [
+export const buttonSizes: Array<ButtonProps['size']> = [
     'xsmall',
     'small-expressive',
     'small-productive',
@@ -18,7 +40,7 @@ export const buttonSizes: ButtonProps['size'][] = [
 /**
  * Button style variants
  */
-export const buttonVariants: ButtonProps['variant'][] = [
+export const buttonVariants: Array<ButtonProps['variant']> = [
     'primary',
     'secondary',
     'outline',
@@ -28,7 +50,7 @@ export const buttonVariants: ButtonProps['variant'][] = [
 /**
  * Button type variants
  */
-export const buttonTypes: ButtonProps['type'][] = [
+export const buttonTypes: Array<ButtonProps['type']> = [
     'submit',
     'button',
     'reset',